Question #12
Doesn't pass the last 2 test cases on hacker rank due to timeout. I
could pre-calculate all of the divisors for the triangle number and I
will probably be able to pass all the cases but that will slow the rest
of the cases down too and I don't like that. I will come back if I
manage to think of a more efficient algo.